work on loading scenes, created editor console

+++ engine/src/core/Font.cpp
@@ -133,7 +133,7 @@ void Font::load(std::string filepath)
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            (unsigned int)face->glyph->advance.x
         };
 
         characters.insert(std::pair<GLchar, Character>(c, character));
